Refactor: image 패키지 구성
